requirements_venv.txt
# aiohappyeyeballs==2.6.1
# aiohttp==3.12.13
# aiosignal==1.3.2
# altair==5.5.0
# annotated-types==0.7.0
# anyio==4.9.0
# argon2-cffi==25.1.0
# argon2-cffi-bindings==21.2.0
# arrow==1.3.0
# asttokens==3.0.0
# async-lru==2.0.5
# attrs==25.3.0
# babel==2.17.0
# backoff==2.2.1
# bcrypt==4.3.0
# beautifulsoup4==4.13.4
# bleach==6.2.0
# blinker==1.9.0
# build==1.2.2.post1
# cachetools==5.5.2
# certifi==2025.6.15
# cffi==1.17.1
# charset-normalizer==3.4.2
# chromadb==1.0.13
# colorama==0.4.6
# coloredlogs==15.0.1
# comm==0.2.2
# contourpy==1.3.2
# cryptography==45.0.4
# cycler==0.12.1
# dataclasses-json==0.6.7
# debugpy==1.8.14
# decorator==5.2.1
# defusedxml==0.7.1
# distro==1.9.0
# dotenv==0.9.9
# durationpy==0.10
# executing==2.2.0
# fastapi
# fastjsonschema==2.21.1
# filelock==3.18.0
# flatbuffers==25.2.10
# fonttools==4.58.4
# fqdn==1.5.1
# frozenlist==1.7.0
# fsspec==2025.5.1
# gitdb==4.0.12
# GitPython==3.1.44
# google-auth==2.40.3
# googleapis-common-protos==1.70.0
# greenlet==3.2.3
# grpcio==1.73.0
# h11==0.16.0
# httpcore==1.0.9
# httptools==0.6.4
# httpx==0.28.1
# httpx-sse==0.4.0
# huggingface-hub==0.33.0
# humanfriendly==10.0
# ibm-cos-sdk==2.13.6
# ibm-cos-sdk-core==2.13.6
# ibm-cos-sdk-s3transfer==2.13.6
# ibm_watson_machine_learning==1.0.367
# ibm_watsonx_ai==1.3.26
# idna==3.10
# importlib_metadata==8.7.0
# importlib_resources==6.5.2
# ipykernel==6.29.5
# ipython==9.3.0
# ipython_pygments_lexers==1.1.1
# ipywidgets==8.1.7
# isoduration==20.11.0
# jedi==0.19.2
# Jinja2==3.1.6
# jmespath==1.0.1
# joblib==1.5.1
# json5==0.12.0
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonschema==4.24.0
# jsonschema-specifications==2025.4.1
# jupyter==1.1.1
# jupyter_client==8.6.3
# jupyter-console==6.6.3
# jupyter_core==5.8.1
# jupyter-events==0.12.0
# jupyter-lsp==2.2.5
# jupyter_server==2.16.0
# jupyter_server_terminals==0.5.3
# jupyterlab==4.4.3
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.27.3
# jupyterlab_widgets==3.0.15
# kiwisolver==1.4.8
# kubernetes==33.1.0
# langchain==0.3.26
# langchain-chroma==0.2.4
# langchain-community==0.3.26
# langchain-core==0.3.66
# langchain-huggingface==0.3.0
# langchain-ibm==0.3.12
# langchain-text-splitters==0.3.8
# langsmith==0.4.1
# lomond==0.3.3
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# marshmallow==3.26.1
# matplotlib==3.10.3
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# mistune==3.1.3
# mmh3==5.1.0
# mpmath==1.3.0
# multidict==6.5.0
# mypy_extensions==1.1.0
# narwhals==1.43.1
# nbclient==0.10.2
# nbconvert==7.16.6
# nbformat==5.10.4
# nest-asyncio==1.6.0
# networkx==3.5
# notebook==7.4.3
# notebook_shim==0.2.4
# numpy==1.26.4
# oauthlib==3.3.1
# onnxruntime==1.22.0
# opentelemetry-api==1.34.1
# opentelemetry-exporter-otlp-proto-common==1.34.1
# opentelemetry-exporter-otlp-proto-grpc==1.34.1
# opentelemetry-proto==1.34.1
# opentelemetry-sdk==1.34.1
# opentelemetry-semantic-conventions==0.55b1
# orjson==3.10.18
# overrides==7.7.0
# packaging==24.2
# pandas==2.1.4
# pandocfilters==1.5.1
# parso==0.8.4
# pdfminer.six==20250506
# pdfplumber==0.11.7
# pillow==11.2.1
# pip==25.1.1
# platformdirs==4.3.8
# plotly==6.1.2
# posthog==5.4.0
# prometheus_client==0.22.1
# prompt_toolkit==3.0.51
# propcache==0.3.2
# protobuf==5.29.5
# psutil==7.0.0
# pure_eval==0.2.3
# pyarrow==20.0.0
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pybase64==1.4.1
# pycparser==2.22
# pydantic==2.11.7
# pydantic_core==2.33.2
# pydantic-settings==2.10.0
# pydeck==0.9.1
# Pygments==2.19.2
# pyparsing==3.2.3
# pypdf==5.6.0
# pypdfium2==4.30.1
# PyPika==0.48.9
# pyproject_hooks==1.2.0
# pyreadline3==3.5.4
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.0
# python-json-logger==3.3.0
# python-multipart
# pytz==2025.2
# pywin32==310
# pywinpty==2.0.15
# PyYAML==6.0.2
# pyzmq==27.0.0
# referencing==0.36.2
# regex==2024.11.6
# requests==2.32.2
# requests-oauthlib==2.0.0
# requests-toolbelt==1.0.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rich==14.0.0
# rpds-py==0.25.1
# rsa==4.9.1
# safetensors==0.5.3
# scikit-learn==1.7.0
# scipy==1.15.3
# Send2Trash==1.8.3
# sentence-transformers==4.1.0
# setuptools==65.5.0
# shellingham==1.5.4
# six==1.17.0
# smmap==5.0.2
# sniffio==1.3.1
# soupsieve==2.7
# SQLAlchemy==2.0.41
# stack-data==0.6.3
# streamlit==1.46.0
# sympy==1.14.0
# tabulate==0.9.0
# tenacity==9.1.2
# terminado==0.18.1
# threadpoolctl==3.6.0
# tinycss2==1.4.0
# tokenizers==0.21.1
# toml==0.10.2
# torch==2.7.1
# tornado==6.5.1
# tqdm==4.67.1
# traitlets==5.14.3
# transformers==4.52.4
# typer==0.16.0
# types-python-dateutil==2.9.0.20250516
# typing_extensions==4.14.0
# typing-inspect==0.9.0
# typing-inspection==0.4.1
# tzdata==2025.2
# uri-template==1.3.0
# urllib3==2.5.0
# uvicorn==0.34.3
# watchdog==6.0.0
# watchfiles==1.1.0
# wcwidth==0.2.13
# webcolors==24.11.1
# webencodings==0.5.1
# websocket-client==1.8.0
# websockets==15.0.1
# widgetsnbextension==4.0.14
# yarl==1.20.1
# zipp==3.23.0
# zstandard==0.23.0
# SpeechRecognition==3.14.3
# streamlit==1.46.0
# pyaudio==0.2.14
# pymysql==1.1.1
# gtts==2.5.4
# pdfkit==1.0.0
# Torchvision==0.22.1

streamlit
supabase
python-dotenv
requests